made changes to the search bar func and delete groups
